+++ 1er semestre/dotnet/Practicas/TrabajoFinal/Aseguradora/Aseguradora.Repositorios/RepositorioVehiculo.cs
@@ -1,5 +1,6 @@
 using Aseguradora.Aplicacion.Interfaces;
 using Aseguradora.Aplicacion.Entidades;
+using Microsoft.EntityFrameworkCore;
 
 namespace Aseguradora.Repositorios;
 
@@ -41,21 +42,51 @@ public class RepositorioVehiculo : IRepositorioVehiculo
     {
         using (var db = new AseguradoraContext())
         {
-            var vehiculoABorrar = db.Vehiculos.Where(v => v.Id == id).SingleOrDefault();
+            // var vehiculoABorrar = db.Vehiculos.Where(v => v.Id == id).SingleOrDefault();
+            // if (vehiculoABorrar != null)
+            // {
+            //     foreach (var poliza in vehiculoABorrar.Polizas)
+            //     {
+            //         foreach (var siniestro in poliza.Siniestros)
+            //         {
+            //             foreach (var tercero in siniestro.Terceros)
+            //             {
+            //                 db.Remove(tercero);
+            //             }
+            //             db.Remove(siniestro);
+            //         }
+            //         db.Remove(poliza);
+            //     }
+            //     db.Remove(vehiculoABorrar);
+            //     db.SaveChanges();
+            // }
+
+            var vehiculoABorrar = db.Vehiculos.Where(v => v.Id == id).Include(v => v.Polizas).SingleOrDefault();
             if (vehiculoABorrar != null)
             {
-                foreach (var poliza in vehiculoABorrar.Polizas)
+                vehiculoABorrar.Polizas?.ToList().
+                ForEach(poliza =>
                 {
-                    foreach (var siniestro in poliza.Siniestros)
+                    var polizaABorrar = db.Polizas.Where(p => p.Id == poliza.Id).Include(p => p.Siniestros).SingleOrDefault();
+                    if (polizaABorrar != null)
                     {
-                        foreach (var tercero in siniestro.Terceros)
+                        polizaABorrar.Siniestros?.ToList().
+                        ForEach(siniestro =>
                         {
-                            db.Remove(tercero);
-                        }
-                        db.Remove(siniestro);
+                            var siniestroABorrar = db.Siniestros.Where(s => s.Id == siniestro.Id).Include(s => s.Terceros).SingleOrDefault();
+                            if (siniestroABorrar != null)
+                            {
+                                siniestroABorrar.Terceros?.ToList().
+                                ForEach(tercero =>
+                                {
+                                    db.Remove(tercero);
+                                });
+                                db.Remove(siniestroABorrar);
+                            }
+                        });
+                        db.Remove(polizaABorrar);
                     }
-                    db.Remove(poliza);
-                }
+                });
                 db.Remove(vehiculoABorrar);
                 db.SaveChanges();
             }
